requirements.txt
pathos
# clize
numpy >= 1.20.0
toml
coverage >= 4.0.0
wheel
setuptools
GitPython